formatter: Use black to keep consistent formating
black [1] is an opinionated python source code formatter.
It improves readability and reduces opinionated review
comments about code format, keeping also a consistent
format across the project.

This change adds `black` to the dependencies and allows
one to run it on the source code to auto-format it:
- Use `black -S ./` in your environment without a container.
- Use `podman run -ti --rm -v $(pwd):/workspace/django-testing-tutorial:Z localhost/beyond black -S ./`
  to run it with a container.

Note: The `-S` option customizes `black` to ignore the string
wrapper character (`'` or `"`). The one used by Django is not
in sync with the opinionated one `black` prefers.

Signed-off-by: Edward Haas <edwardh@redhat.com>